Implement config hook in addon's index.js
This allows for the addon to specify any settings it needs enabled in
the consuming app.

Also update the package.json (aplogies for the reformat). I had to move
the non-core-Ember npm dependencies from dev to regular dependencies. It
appears the consuming app will not install dev deps.

Also add blueprints for .env files.

Update the README to reflect this all!