Fix error where cookie auth beforeModel hook failed
When user was not logged in, beforeModel hook returned a rejected promise and failed. Catch that error and do nothing, so hook can pass through.
(it just

BeforeModel hook (and ember simple auth) expect an ember promise, but receives a jquery object. Massage the promise type so things work correctly.

[#PREP-87]