Remove nice-to-have regression test because it doesn't work
Making this test work will require mocking out some calls, which is tricky because we have multiple uses of a method wrapped inside a single save() call. When we stub contributor.save completely, the test passes even on develop (because we're bypassing the very store manipulation that was causing the regression. This will require a different approach to how/what we mock, and contributors have some unique save rules layered on.

Suggested future regression test:
- Add a contributor
- Then edit the contributor's permissions and commit the data. (in the OSF contribs page, this calls node.save). As of 8c397df18bd2f444c9f6af2ded03577ce22297c1, this should fail, but the same test should be expected to pass with this PR in place.

The failure was evidenced in the browser address bar, shown by a JS console error about `s.record` not being found.

[#EOSF-136]